requirements.txt
pyspark==3.3.1
pyyaml==5.4.1
delta-spark==2.1.0
pytest